requirements.txt
python-dotenv==1.0.0
django==5.1.1
django-types==0.19.1
djangorestframework==3.15.2
djangorestframework-types==0.8.0
django-filter==24.3
psycopg2-binary==2.9.9
gunicorn==23.0.0
requests==2.32.3
types-requests==2.32.0.20240914
pydantic==2.9.2
redis==5.1.0
flake8==7.1.1
pylint==3.3.1
mypy==1.11.2
